dissipation model testing

+++ src/DissipationModelRandom.cxx
@@ -1,4 +1,5 @@
 #include "DissipationModelRandom.h"
+#include <cstddef>
 
 DissipationModelRandom::DissipationModelRandom(){
     this->rangeMin = 0;
@@ -15,7 +16,7 @@ DissipationModelRandom::~DissipationModelRandom(){
 
 arma::Col<double> DissipationModelRandom::dissipate(arma::Col<double> input, double time){
     arma::Col<double> output = arma::Col<double>(input.n_rows);
-    for(int i = 0; i < input.n_rows; i++){
+    for(size_t i = 0; i < input.n_rows; i++){
         output(i) = input(i) - input(i)*(this->rangeMin + static_cast <float> (rand()) /( static_cast <float> (RAND_MAX/(this->rangeMax-this->rangeMin))));
     }
     return output;
@@ -23,7 +24,7 @@ arma::Col<double> DissipationModelRandom::dissipate(arma::Col<double> input, dou
 
 arma::Col<double> DissipationModelRandom::dissipateTerm(arma::Col<double> input, double time){
     arma::Col<double> output = arma::Col<double>(input.n_rows);
-    for(int i = 0; i < input.n_rows; i++){
+    for(size_t i = 0; i < input.n_rows; i++){
         output(i) = input(i)*(this->rangeMin + static_cast <float> (rand()) /( static_cast <float> (RAND_MAX/(this->rangeMax-this->rangeMin))));
     }
     return output;
